+++ core/openapi/source/org/jetbrains/mps/openapi/language/SPropertyId.java
@@ -59,7 +59,7 @@ public final class SPropertyId {
   public static SPropertyId deserialize(String s) {
     int split = s.lastIndexOf("/");
     SConceptId concept = SConceptId.deserialize(s.substring(0, split));
-    int prop = Integer.parseInt(s.substring(split + 1));
+    long prop = Long.parseLong(s.substring(split + 1));
     return new SPropertyId(concept, prop);
   }
 }
