+++ core/openapi/source/org/jetbrains/mps/openapi/language/SReferenceLinkId.java
@@ -54,7 +54,7 @@ public final class SReferenceLinkId extends SAbstractLinkId {
   public static SReferenceLinkId deserialize(String s) {
     int split = s.lastIndexOf("/");
     SConceptId concept = SConceptId.deserialize(s.substring(0, split));
-    int ref = Integer.parseInt(s.substring(split + 1));
+    long ref = Long.parseLong(s.substring(split + 1));
     return new SReferenceLinkId(concept, ref);
   }
 }
